fix: address PR review mergeability blockers
Move provider-session detail HTTP tests out of server_factory_test.go to satisfy backend-size, and align CLI baseline fixtures with the current you-agent-factories global root layout.
