+++ pkg/cli/goal_failure_baseline_test.go
@@ -487,7 +487,7 @@ func TestFailureBaseline_NamedPath_RunNamedGoalSurfacesPercentEncodedFactoryDir(
 	if filepath.Base(got.Dir) != "@you%2Fgoal" {
 		t.Fatalf("run dir base = %q, want @you%%2Fgoal layout segment", filepath.Base(got.Dir))
 	}
-	if !strings.Contains(got.Dir, filepath.Join(".you-agent-factory", "factories")) {
+	if !strings.Contains(got.Dir, filepath.Join(".you-agent-factory", "you-agent-factories")) {
 		t.Fatalf("run dir = %q, want global named-factory root layout", got.Dir)
 	}
 }
